Add a simple benchmark timing the collect-authority example
In order to make scenario execution benchmarkable, we also need to
split the crate into a library and a binary.
