+++ Cargo.toml
@@ -25,3 +25,10 @@ zip = "0.5.5"
 
 [build-dependencies]
 protobuf-codegen-pure = "2.14.0"
+
+[dev-dependencies]
+criterion = "0.3.2"
+
+[[bench]]
+name = "scenario"
+harness = false
