+++ NOTICES
@@ -1,46 +1,98 @@
 adler32: 1.0.3, "BSD-3-Clause AND Zlib",
 aho-corasick: 0.7.10, "MIT OR Unlicense",
+atty: 0.2.14, "MIT",
 autocfg: 1.0.0, "Apache-2.0 OR MIT",
 bigdecimal: 0.1.2, "Apache-2.0 OR MIT",
+bitflags: 1.2.1, "Apache-2.0 OR MIT",
 block-buffer: 0.7.3, "Apache-2.0 OR MIT",
 block-padding: 0.1.4, "Apache-2.0 OR MIT",
+bstr: 0.2.12, "Apache-2.0 OR MIT",
+bumpalo: 3.2.1, "Apache-2.0 OR MIT",
 byte-tools: 0.3.1, "Apache-2.0 OR MIT",
 byteorder: 1.3.2, "MIT OR Unlicense",
 bzip2-sys: 0.1.7, "Apache-2.0 OR MIT",
 bzip2: 0.3.3, "Apache-2.0 OR MIT",
+cast: 0.2.3, "Apache-2.0 OR MIT",
 cc: 1.0.37, "Apache-2.0 OR MIT",
-cfg-if: 0.1.6, "Apache-2.0 OR MIT",
+cfg-if: 0.1.10, "Apache-2.0 OR MIT",
+clap: 2.33.0, "MIT",
 crc32fast: 1.2.0, "Apache-2.0 OR MIT",
+criterion-plot: 0.4.2, "Apache-2.0 OR MIT",
+criterion: 0.3.2, "Apache-2.0 OR MIT",
+crossbeam-deque: 0.7.3, "Apache-2.0 OR MIT",
+crossbeam-epoch: 0.8.2, "Apache-2.0 OR MIT",
+crossbeam-queue: 0.2.1, "Apache-2.0 AND BSD-2-Clause OR MIT",
+crossbeam-utils: 0.7.2, "Apache-2.0 OR MIT",
+csv-core: 0.1.10, "MIT OR Unlicense",
+csv: 1.1.3, "MIT OR Unlicense",
 digest: 0.8.0, "Apache-2.0 OR MIT",
+either: 1.5.3, "Apache-2.0 OR MIT",
 fake-simd: 0.1.2, "Apache-2.0 OR MIT",
 flate2: 1.0.14, "Apache-2.0 OR MIT",
 fnv: 1.0.6, "Apache-2.0 OR MIT",
 generic-array: 0.12.0, "MIT",
+hermit-abi: 0.1.12, "Apache-2.0 OR MIT",
 hex: 0.4.2, "Apache-2.0 OR MIT",
+itertools: 0.9.0, "Apache-2.0 OR MIT",
+itoa: 0.4.5, "Apache-2.0 OR MIT",
+js-sys: 0.3.39, "Apache-2.0 OR MIT",
 lazy_static: 1.4.0, "Apache-2.0 OR MIT",
 libc: 0.2.69, "Apache-2.0 OR MIT",
 libflate: 0.1.26, "MIT",
+log: 0.4.8, "Apache-2.0 OR MIT",
+maybe-uninit: 2.0.0, "Apache-2.0 OR MIT",
 memchr: 2.3.3, "MIT OR Unlicense",
+memoffset: 0.5.4, "MIT",
 miniz_oxide: 0.3.5, "MIT",
 num-bigint: 0.2.6, "Apache-2.0 OR MIT",
 num-integer: 0.1.42, "Apache-2.0 OR MIT",
 num-traits: 0.2.11, "Apache-2.0 OR MIT",
+num_cpus: 1.13.0, "Apache-2.0 OR MIT",
+oorandom: 11.1.1, "MIT",
 opaque-debug: 0.2.2, "Apache-2.0 OR MIT",
+plotters: 0.2.12, "MIT",
 podio: 0.1.6, "Apache-2.0 OR MIT",
+proc-macro2: 1.0.11, "Apache-2.0 OR MIT",
 protobuf-codegen-pure: 2.14.0, "MIT",
 protobuf-codegen: 2.14.0, "MIT",
 protobuf: 2.14.0, "MIT",
+quote: 1.0.4, "Apache-2.0 OR MIT",
+rayon-core: 1.7.0, "Apache-2.0 OR MIT",
+rayon: 1.3.0, "Apache-2.0 OR MIT",
 redox_syscall: 0.1.54, "MIT",
+regex-automata: 0.1.9, "MIT OR Unlicense",
 regex-syntax: 0.6.17, "Apache-2.0 OR MIT",
 regex: 1.3.7, "Apache-2.0 OR MIT",
 rle-decode-fast: 1.0.1, "Apache-2.0 OR MIT",
+rustc_version: 0.2.3, "Apache-2.0 OR MIT",
 rusty-engine: 0.0.1, "Apache-2.0",
+ryu: 1.0.4, "Apache-2.0 OR BSL-1.0",
+same-file: 1.0.6, "MIT OR Unlicense",
+scopeguard: 1.1.0, "Apache-2.0 OR MIT",
+semver-parser: 0.7.0, "Apache-2.0 OR MIT",
+semver: 0.9.0, "Apache-2.0 OR MIT",
+serde: 1.0.106, "Apache-2.0 OR MIT",
+serde_derive: 1.0.106, "Apache-2.0 OR MIT",
+serde_json: 1.0.52, "Apache-2.0 OR MIT",
 sha2: 0.8.1, "Apache-2.0 OR MIT",
+syn: 1.0.18, "Apache-2.0 OR MIT",
 take_mut: 0.2.2, "MIT",
+textwrap: 0.11.0, "MIT",
 thread_local: 1.0.1, "Apache-2.0 OR MIT",
 time: 0.1.42, "Apache-2.0 OR MIT",
+tinytemplate: 1.0.4, "Apache-2.0 OR MIT",
 typenum: 1.10.0, "Apache-2.0 OR MIT",
+unicode-width: 0.1.7, "Apache-2.0 OR MIT",
+unicode-xid: 0.2.0, "Apache-2.0 OR MIT",
+walkdir: 2.3.1, "MIT OR Unlicense",
+wasm-bindgen-backend: 0.2.62, "Apache-2.0 OR MIT",
+wasm-bindgen-macro-support: 0.2.62, "Apache-2.0 OR MIT",
+wasm-bindgen-macro: 0.2.62, "Apache-2.0 OR MIT",
+wasm-bindgen-shared: 0.2.62, "Apache-2.0 OR MIT",
+wasm-bindgen: 0.2.62, "Apache-2.0 OR MIT",
+web-sys: 0.3.39, "Apache-2.0 OR MIT",
 winapi-i686-pc-windows-gnu: 0.4.0, "Apache-2.0 OR MIT",
+winapi-util: 0.1.5, "MIT OR Unlicense",
 winapi-x86_64-pc-windows-gnu: 0.4.0, "Apache-2.0 OR MIT",
 winapi: 0.3.6, "Apache-2.0 OR MIT",
 zip: 0.5.5, "MIT",
