+++ test/collect-authority/daml.yaml
@@ -1,0 +1,7 @@
+sdk-version: 1.0.0
+name: collect-authority
+source: CollectAuthority.daml
+version: 0.0.1
+dependencies:
+  - daml-prim
+  - daml-stdlib
